fix jwt logic and add realt CRUD

+++ src/main/java/com/yashny/realestate_backend/services/RealtService.java
@@ -2,10 +2,7 @@ package com.yashny.realestate_backend.services;
 
 import com.yashny.realestate_backend.entities.Realt;
 import com.yashny.realestate_backend.entities.User;
-import com.yashny.realestate_backend.repositories.DealTypeRepository;
 import com.yashny.realestate_backend.repositories.RealtRepository;
-import com.yashny.realestate_backend.repositories.TypeRepository;
-import com.yashny.realestate_backend.repositories.UserRepository;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -14,7 +11,7 @@ import org.springframework.web.multipart.MultipartFile;
 import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -41,11 +38,16 @@ public class RealtService {
         realtRepository.save(realt);
     }
 
-    public void deleteRealt(Long id) {
+    public boolean deleteRealt(Long id, User user) {
         Realt realt = realtRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Realt not found with id " + id));
 
-        realtRepository.delete(realt);
+        if (realt.getUser() == user || Objects.equals(user.getRole(), "ADMIN") || Objects.equals(user.getRole(), "SUPER_ADMIN")) {
+            realtRepository.delete(realt);
+            return true;
+        }
+
+        return false;
     }
 
     public List<String> uploadImages(MultipartFile[] files) throws Exception {
